Add Checkout function, Change path ui/home -> /home, Add is_delivered column into orders table, Import UTM font

+++ database/migrations/2018_02_28_021807_create_users_table.php
@@ -20,7 +20,7 @@ class CreateUsersTable extends Migration
             $table->string('password');
             $table->string('phone', 11)->nullable();
             $table->integer('role_id')->unsigned()->default(1);
-            $table->integer('is_active')->default(1);
+            $table->tinyInteger('is_active')->default(1);
             $table->rememberToken();
             $table->timestamps();
         });
